pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-researcher"
version = "0.1.0"
description = "Let AI agents run experiments in any repo while you sleep."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Open Researcher Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]
keywords = ["ai", "research", "experiments", "automation", "agent"]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "textual>=0.85.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[project.scripts]
open-researcher = "open_researcher.cli:app"

[project.urls]
Homepage = "https://github.com/open-researcher/open-researcher"
Repository = "https://github.com/open-researcher/open-researcher"
Issues = "https://github.com/open-researcher/open-researcher/issues"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
